# docs/swagger.yaml
basePath: /
definitions:
  github_com_0xPexy_sentra-backend_internal_admin.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  github_com_0xPexy_sentra-backend_internal_indexer_service.AssetMovement:
    properties:
      address:
        type: string
      delta:
        type: string
      token:
        type: string
      tokenId:
        type: string
    type: object
  github_com_0xPexy_sentra-backend_internal_indexer_service.GasBreakdown:
    properties:
      from:
        type: string
      gas:
        type: string
      gasUsed:
        type: string
      method:
        type: string
      stage:
        type: string
      to:
        type: string
      type:
        type: string
    type: object
...
  github_com_0xPexy_sentra-backend_internal_indexer_service.NFTTokenItem:
    properties:
      contract:
        type: string
      owner:
        type: string
      tokenId:
        type: string
    type: object
  github_com_0xPexy_sentra-backend_internal_indexer_service.PaymasterOverview:
    properties:
      avgActualGasUsed:
        type: number
      successRate:
        type: number
      totalSponsoredGasCost:
        type: string
      totalSponsoredOps:
        type: integer
    type: object
  github_com_0xPexy_sentra-backend_internal_indexer_service.PhaseGas:
    properties:
      gasLimit:
        type: string
      gasUsed:
        type: string
      phase:
        type: string
    type: object
  github_com_0xPexy_sentra-backend_internal_indexer_service.RevertInfo:
    properties:
      message:
        type: string
      raw:
        type: string
      selector:
        type: string
    type: object
  github_com_0xPexy_sentra-backend_internal_indexer_service.SponsoredOpItem:
    properties:
      blockNumber:
        type: integer
      blockTime:
        type: string
      logIndex:
        type: integer
      selector:
        type: string
      sender:
        type: string
      status:
        type: string
      target:
        type: string
      userOpHash:
        type: string
    type: object
  github_com_0xPexy_sentra-backend_internal_indexer_service.SponsoredOpsResult:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_0xPexy_sentra-backend_internal_indexer_service.SponsoredOpItem'
        type: array
      nextCursor:
        type: string
    type: object
  github_com_0xPexy_sentra-backend_internal_indexer_service.SponsorshipInfo:
    properties:
      validAfter:
        type: string
      validUntil:
        type: string
    type: object
  github_com_0xPexy_sentra-backend_internal_indexer_service.UserOpEventInfo:
    properties:
      blockNumber:
        type: integer
      name:
        type: string
      txHash:
        type: string
    type: object
  github_com_0xPexy_sentra-backend_internal_indexer_service.UserOperationDetail:
    properties:
      actualGasCost:
        type: string
      actualGasUsed:
        type: string
      assetMovements:
        items:
          $ref: '#/definitions/github_com_0xPexy_sentra-backend_internal_indexer_service.AssetMovement'
        type: array
      beneficiary:
        type: string
      blockNumber:
        type: integer
      blockTime:
        type: string
      callGasLimit:
        type: string
      events:
        items:
          $ref: '#/definitions/github_com_0xPexy_sentra-backend_internal_indexer_service.UserOpEventInfo'
        type: array
      gasBreakdown:
        items:
          $ref: '#/definitions/github_com_0xPexy_sentra-backend_internal_indexer_service.GasBreakdown'
        type: array
      logIndex:
        type: integer
      maxFeePerGas:
        type: string
      maxPriorityFeePerGas:
        type: string
      mintedNfts:
        items:
          $ref: '#/definitions/github_com_0xPexy_sentra-backend_internal_indexer_service.NFTTokenItem'
        type: array
      nonce:
        type: string
      paymaster:
        type: string
      paymasterPostOpGasLimit:
        type: string
      paymasterVerificationGasLimit:
        type: string
      preVerificationGas:
        type: string
      revert:
        $ref: '#/definitions/github_com_0xPexy_sentra-backend_internal_indexer_service.RevertInfo'
      revertReason:
        type: string
      selector:
        type: string
      sender:
        type: string
      sponsorship:
        $ref: '#/definitions/github_com_0xPexy_sentra-backend_internal_indexer_service.SponsorshipInfo'
      status:
        type: string
      target:
        type: string
      txHash:
        type: string
      userOpHash:
        type: string
      verificationGasLimit:
        type: string
    type: object
  github_com_0xPexy_sentra-backend_internal_indexer_service.UserOperationGas:
    properties:
      actualGasCost:
        type: string
      actualGasUsed:
        type: string
      callGasLimit:
        type: string
      maxFeePerGas:
        type: string
      maxPriorityFeePerGas:
        type: string
      paymasterPostOpGasLimit:
        type: string
      paymasterVerificationGasLimit:
        type: string
      phases:
        items:
          $ref: '#/definitions/github_com_0xPexy_sentra-backend_internal_indexer_service.PhaseGas'
        type: array
      preVerificationGas:
        type: string
      txHash:
        type: string
      userOpHash:
        type: string
      verificationGasLimit:
        type: string
    type: object
...
info:
  contact: {}
  description: API documentation for the Sentinel 4337 backend service.
  title: Sentinel 4337 Backend API
  version: "1.0"
paths:
...
  /api/v1/ops/{userOpHash}:
    get:
      parameters:
      - description: User operation hash
        in: path
        name: userOpHash
        required: true
        type: string
      - description: Chain ID override
        format: int64
        in: query
        name: chain_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_0xPexy_sentra-backend_internal_indexer_service.UserOperationDetail'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_0xPexy_sentra-backend_internal_admin.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_0xPexy_sentra-backend_internal_admin.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_0xPexy_sentra-backend_internal_admin.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user operation detail
      tags:
      - Operations
  /api/v1/ops/{userOpHash}/gas:
    get:
      parameters:
      - description: User operation hash
        in: path
        name: userOpHash
        required: true
        type: string
      - description: Chain ID override
        format: int64
        in: query
        name: chain_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_0xPexy_sentra-backend_internal_indexer_service.UserOperationGas'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_0xPexy_sentra-backend_internal_admin.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_0xPexy_sentra-backend_internal_admin.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_0xPexy_sentra-backend_internal_admin.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user operation gas breakdown
      tags:
      - Operations
...
  /api/v1/paymasters/ops:
    get:
      parameters:
      - description: Chain ID override
        format: int64
        in: query
        name: chain_id
        type: integer
      - description: Page size (1-100)
        in: query
        name: limit
        type: integer
      - description: Cursor in blockNumber:logIndex format
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_0xPexy_sentra-backend_internal_indexer_service.SponsoredOpsResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_0xPexy_sentra-backend_internal_admin.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_0xPexy_sentra-backend_internal_admin.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List sponsored user operations for a paymaster
      tags:
      - Operations
...
  /api/v1/stats/overview:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_0xPexy_sentra-backend_internal_indexer_service.PaymasterOverview'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_0xPexy_sentra-backend_internal_admin.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_0xPexy_sentra-backend_internal_admin.ErrorResponse'
      summary: Paymaster stats overview
      tags:
      - Stats
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"